+++ ldscript.txt
@@ -106,6 +106,7 @@ SECTIONS
         build/sa2/asm/interactable_music_plant_guitar_string.o(.text);
         build/sa2/src/interactable_music_plant_keyboard.o(.text);
         build/sa2/src/interactable_music_plant_german_flute.o(.text);
+        build/sa2/src/handlers_ia_pipe_horn.o(.text);
         build/sa2/src/interactable_music_plant_pipe_horn.o(.text);
         build/sa2/src/interactable_ice_paradise_slowing_snow.o(.text);
         build/sa2/asm/interactable_ice_paradise_funnel_sphere.o(.text);
@@ -237,6 +238,7 @@ SECTIONS
         build/sa2/src/interactable_music_plant_note_block.o(.rodata);
         build/sa2/src/interactable_music_plant_keyboard.o(.rodata);
         build/sa2/src/interactable_music_plant_german_flute.o(.rodata);
+        build/sa2/src/handlers_ia_pipe_horn.o(.rodata);
         build/sa2/data/data1_2_2.o(.rodata);
         build/sa2/src/interactable_music_plant_pipe_horn.o(.rodata);
         build/sa2/data/data1_2_3.o(.rodata);
@@ -539,7 +541,7 @@ SECTIONS
         /* Asset data */
         build/sa2/data/data3.o(.rodata);
         build/sa2/data/special_stage_levels.o(.rodata);
-        build/sa2/src/interactable_music_plant_pipe_horn.o(.data);
+        build/sa2/src/handlers_ia_pipe_horn.o(.data);
         build/sa2/src/palettes.o(.rodata);
         build/sa2/src/recordings.o(.rodata);
         build/sa2/data/data4.o(.rodata);
