+++ src/interactable_music_plant_pipe_horn.c
@@ -5,35 +5,10 @@
 #include "interactable.h"
 #include "sprite.h"
 
+#include "interactable_music_plant_pipe_horn.h"
+
 #include "constants/move_states.h"
 #include "constants/songs.h"
-
-typedef struct {
-    /* 0x00 */ s32 x0;
-    /* 0x04 */ s32 y0;
-    /* 0x08 */ s32 x1;
-    /* 0x0C */ s32 y1;
-    /* 0x10 */ s32 x2;
-    /* 0x14 */ s32 y2;
-    /* 0x18 */ u16 unk18;
-    /* 0x1A */ u16 unk1A;
-    /* 0x1C */ s32 posX;
-    /* 0x20 */ s32 posY;
-    /* 0x24 */ u16 kind;
-    /* 0x26 */ u16 unk26;
-    /* 0x28 */ Interactable *ia;
-    /* 0x2C */ u8 spriteX;
-    /* 0x2D */ u8 spriteY;
-} Sprite_Pipe_Horn;
-
-typedef struct {
-    u16 unk0;
-    u16 unk2;
-    union {
-        s16 _16[2];
-        s32 _32; // Q_24_8?
-    } unk4;
-} Pipe_Data;
 
 static const s16 gUnknown_080DFFF4[3][2] = {
     { Q_8_8(9), Q_8_8(0) },
@@ -63,20 +38,10 @@ extern const Pipe_Data gUnknown_080DFF9C[];
 
 // NOTE: Gets stored in .data section
 //       (I assume that is because it's a pointer array)
-const Pipe_Data (*gUnknown_08C8793C[9])[] = {
-    &gUnknown_080DFCF0, &gUnknown_080DFCF0, &gUnknown_080DFD40,
-    &gUnknown_080DFD40, &gUnknown_080DFD98, &gUnknown_080DFDD8,
-    &gUnknown_080DFDD8, &gUnknown_080DFE30, &gUnknown_080DFE30,
-};
+extern const Pipe_Data (*gUnknown_08C8793C[9])[];
+extern const Pipe_Data (*gUnknown_08C87960[3])[];
 
-// NOTE: Gets stored in .data section
-const Pipe_Data (*gUnknown_08C87960[3])[] = {
-    &gUnknown_080DFEE4,
-    &gUnknown_080DFF3C,
-    &gUnknown_080DFF9C,
-};
-
-extern void (*sPipeHandlers[16])(Sprite_Pipe_Horn *, const Pipe_Data data[]);
+extern PipeHandler sPipeHandlers[NUM_PIPE_HANDLERS];
 
 extern void Player_SetMovestate_IsInScriptedSequence(void);
 extern void Player_ClearMovestate_IsInScriptedSequence(void);
@@ -102,370 +67,6 @@ extern bool32 sub_8077CB0(Sprite_Pipe_Horn *);
 extern void Task_FrenchHorn_8077C04(void);
 void FrenchHorn_Despawn(Sprite_Pipe_Horn *);
 void TaskDestructor_FrenchHorn(struct Task *);
-
-void Handler_MusicPlant_Pipe_0(Sprite_Pipe_Horn *pipe, const Pipe_Data data[])
-{
-    s32 r1 = (u16)data[pipe->unk18].unk4._16[1];
-    s32 r0, r3, r6, r7;
-    s16 *data2 = data[pipe->unk18].unk4._16;
-
-    r0 = data2[0];
-    r3 = pipe->x1;
-    r7 = -r0;
-    pipe->x2 = r3 - (((s16)pipe->unk1A * r7) >> 10);
-
-#ifndef NON_MATCHING
-    asm("lsl %0, %0, #16\n"
-        "asr %0, %0, #16"
-        :
-        : "r"(r1));
-#else
-    r1 = (s16)r1;
-#endif
-
-    r6 = pipe->y1;
-    pipe->y2 = r6 - ((-r1 * (s16)pipe->unk1A) >> 10);
-
-    pipe->unk1A += data[pipe->unk18].unk2;
-
-    if (pipe->unk1A > ONE_CYCLE) {
-        u32 value;
-        pipe->unk1A -= 1024;
-
-        pipe->x1 = pipe->x1 - r7;
-        pipe->y1 = pipe->y1 - -r1;
-
-        value = data[++pipe->unk18].unk0;
-        if (value == (u16)-1) {
-            pipe->unk18 |= value;
-        }
-    }
-}
-
-void Handler_MusicPlant_Pipe_1(Sprite_Pipe_Horn *pipe, const Pipe_Data data[])
-{
-    s16 sin, cos;
-    s32 r5 = data[pipe->unk18].unk4._32;
-    u32 index = (pipe->unk1A >> 2);
-    u16 sinIndex = (DEG_TO_SIN(90) - index) & ONE_CYCLE;
-
-    cos = COS_24_8(sinIndex);
-    pipe->x2 = Q_24_8_TO_INT(cos * r5) + pipe->x1;
-
-    sin = SIN_24_8(sinIndex);
-    pipe->y2 = Q_24_8_TO_INT(sin * r5) + pipe->y1 - r5;
-
-    pipe->unk1A += data[pipe->unk18].unk2;
-
-    if (pipe->unk1A > ONE_CYCLE) {
-        u32 value;
-        pipe->unk1A -= 1024;
-
-        cos = COS_DEG(0);
-        pipe->x1 = Q_24_8_TO_INT(cos * r5) + pipe->x1;
-
-        sin = SIN_DEG(0);
-        pipe->y1 = Q_24_8_TO_INT(sin * r5) + pipe->y1 - r5;
-
-        value = data[++pipe->unk18].unk0;
-        if (value == (u16)-1) {
-            pipe->unk18 |= value;
-        }
-    }
-}
-
-void Handler_MusicPlant_Pipe_2(Sprite_Pipe_Horn *pipe, const Pipe_Data data[])
-{
-    s16 sin, cos;
-    s32 r5 = data[pipe->unk18].unk4._32;
-    u32 index = (pipe->unk1A >> 2);
-    u16 sinIndex = (DEG_TO_SIN(270) + index) & ONE_CYCLE;
-
-    cos = COS_24_8(sinIndex);
-    pipe->x2 = Q_24_8_TO_INT(cos * r5) + pipe->x1;
-
-    sin = SIN_24_8(sinIndex);
-    pipe->y2 = Q_24_8_TO_INT(sin * r5) + pipe->y1 + r5;
-
-    pipe->unk1A += data[pipe->unk18].unk2;
-
-    if (pipe->unk1A > ONE_CYCLE) {
-        u32 value;
-        pipe->unk1A -= 1024;
-
-        cos = COS_DEG(0);
-        pipe->x1 = Q_24_8_TO_INT(cos * r5) + pipe->x1;
-
-        sin = SIN_DEG(0);
-        pipe->y1 = Q_24_8_TO_INT(sin * r5) + pipe->y1 + r5;
-
-        value = data[++pipe->unk18].unk0;
-        if (value == (u16)-1) {
-            pipe->unk18 |= value;
-        }
-    }
-}
-
-void Handler_MusicPlant_Pipe_3(Sprite_Pipe_Horn *pipe, const Pipe_Data data[])
-{
-    s16 sin, cos;
-    s32 r5 = data[pipe->unk18].unk4._32;
-    u32 index = (pipe->unk1A >> 2);
-    u16 sinIndex = (index)&ONE_CYCLE;
-
-    cos = COS_24_8(sinIndex);
-    pipe->x2 = Q_24_8_TO_INT(cos * r5) + pipe->x1 - r5;
-
-    sin = SIN_24_8(sinIndex);
-    pipe->y2 = Q_24_8_TO_INT(sin * r5) + pipe->y1;
-
-    pipe->unk1A += data[pipe->unk18].unk2;
-
-    if (pipe->unk1A > ONE_CYCLE) {
-        u32 value;
-        pipe->unk1A -= 1024;
-
-        cos = COS_DEG(90);
-        pipe->x1 = Q_24_8_TO_INT(cos * r5) + pipe->x1 - r5;
-
-        sin = SIN_DEG(90);
-        pipe->y1 = Q_24_8_TO_INT(sin * r5) + pipe->y1;
-
-        value = data[++pipe->unk18].unk0;
-        if (value == (u16)-1) {
-            pipe->unk18 |= value;
-        }
-    }
-}
-
-void Handler_MusicPlant_Pipe_4(Sprite_Pipe_Horn *pipe, const Pipe_Data data[])
-{
-    s16 sin, cos;
-    s32 r5 = data[pipe->unk18].unk4._32;
-    u32 index = (pipe->unk1A >> 2);
-    u16 sinIndex = (DEG_TO_SIN(180) - index) & ONE_CYCLE;
-
-    cos = COS_24_8(sinIndex);
-    pipe->x2 = Q_24_8_TO_INT(cos * r5) + pipe->x1 + r5;
-
-    sin = SIN_24_8(sinIndex);
-    pipe->y2 = Q_24_8_TO_INT(sin * r5) + pipe->y1;
-
-    pipe->unk1A += data[pipe->unk18].unk2;
-
-    if (pipe->unk1A > ONE_CYCLE) {
-        u32 value;
-        pipe->unk1A -= 1024;
-
-        cos = COS_DEG(90);
-        pipe->x1 = Q_24_8_TO_INT(cos * r5) + pipe->x1 + r5;
-
-        sin = SIN_DEG(90);
-        pipe->y1 = Q_24_8_TO_INT(sin * r5) + pipe->y1;
-
-        value = data[++pipe->unk18].unk0;
-        if (value == (u16)-1) {
-            pipe->unk18 |= value;
-        }
-    }
-}
-
-void Handler_MusicPlant_Pipe_5(Sprite_Pipe_Horn *pipe, const Pipe_Data data[])
-{
-    s16 sin, cos;
-    s32 r5 = data[pipe->unk18].unk4._32;
-    u32 index = (pipe->unk1A >> 2);
-    u16 sinIndex = (index + DEG_TO_SIN(90)) & ONE_CYCLE;
-
-    cos = COS_24_8(sinIndex);
-    pipe->x2 = Q_24_8_TO_INT(cos * r5) + pipe->x1;
-
-    sin = SIN_24_8(sinIndex);
-    pipe->y2 = Q_24_8_TO_INT(sin * r5) + pipe->y1 - r5;
-
-    pipe->unk1A += data[pipe->unk18].unk2;
-
-    if (pipe->unk1A > ONE_CYCLE) {
-        u32 value;
-        pipe->unk1A -= 1024;
-
-        cos = COS_DEG(180);
-        pipe->x1 = Q_24_8_TO_INT(cos * r5) + pipe->x1;
-
-        sin = SIN_DEG(180);
-        pipe->y1 = Q_24_8_TO_INT(sin * r5) + pipe->y1 - r5;
-
-        value = data[++pipe->unk18].unk0;
-        if (value == (u16)-1) {
-            pipe->unk18 |= value;
-        }
-    }
-}
-
-void Handler_MusicPlant_Pipe_6(Sprite_Pipe_Horn *pipe, const Pipe_Data data[])
-{
-    s16 sin, cos;
-    s32 r5 = data[pipe->unk18].unk4._32;
-    u32 index = (pipe->unk1A >> 2);
-    u16 sinIndex = (DEG_TO_SIN(270) - index) & ONE_CYCLE;
-
-    cos = COS_24_8(sinIndex);
-    pipe->x2 = Q_24_8_TO_INT(cos * r5) + pipe->x1;
-
-    sin = SIN_24_8(sinIndex);
-    pipe->y2 = Q_24_8_TO_INT(sin * r5) + pipe->y1 + r5;
-
-    pipe->unk1A += data[pipe->unk18].unk2;
-
-    if (pipe->unk1A > ONE_CYCLE) {
-        u32 value;
-        pipe->unk1A -= 1024;
-
-        cos = COS_DEG(180);
-        pipe->x1 = Q_24_8_TO_INT(cos * r5) + pipe->x1;
-
-        sin = SIN_DEG(180);
-        pipe->y1 = Q_24_8_TO_INT(sin * r5) + pipe->y1 + r5;
-
-        value = data[++pipe->unk18].unk0;
-        if (value == (u16)-1) {
-            pipe->unk18 |= value;
-        }
-    }
-}
-
-void Handler_MusicPlant_Pipe_7(Sprite_Pipe_Horn *pipe, const Pipe_Data data[])
-{
-    s16 sin, cos;
-    s32 r5 = data[pipe->unk18].unk4._32;
-    u32 index = (pipe->unk1A >> 2);
-    u16 sinIndex = (DEG_TO_SIN(360) - index) & ONE_CYCLE;
-
-    cos = COS_24_8(sinIndex);
-    pipe->x2 = Q_24_8_TO_INT(cos * r5) + pipe->x1 - r5;
-
-    sin = SIN_24_8(sinIndex);
-    pipe->y2 = Q_24_8_TO_INT(sin * r5) + pipe->y1;
-
-    pipe->unk1A += data[pipe->unk18].unk2;
-
-    if (pipe->unk1A > ONE_CYCLE) {
-        u32 value;
-        pipe->unk1A -= 1024;
-
-        cos = COS_DEG(270);
-        pipe->x1 = Q_24_8_TO_INT(cos * r5) + pipe->x1;
-        pipe->x1 -= r5;
-
-        sin = SIN_DEG(270);
-        pipe->y1 = Q_24_8_TO_INT(sin * r5) + pipe->y1;
-
-        value = data[++pipe->unk18].unk0;
-        if (value == (u16)-1) {
-            pipe->unk18 |= value;
-        }
-    }
-}
-
-void Handler_MusicPlant_Pipe_8(Sprite_Pipe_Horn *pipe, const Pipe_Data data[])
-{
-    s16 sin, cos;
-    s32 r5 = data[pipe->unk18].unk4._32;
-    u16 sinIndex = ((pipe->unk1A >> 2) + 512) & ONE_CYCLE;
-
-    cos = COS_24_8(sinIndex);
-    pipe->x2 = Q_24_8_TO_INT(cos * r5) + pipe->x1 + r5;
-
-    sin = SIN_24_8(sinIndex);
-    pipe->y2 = Q_24_8_TO_INT(sin * r5) + pipe->y1;
-
-    pipe->unk1A += data[pipe->unk18].unk2;
-
-    if (pipe->unk1A > ONE_CYCLE) {
-        u32 value;
-        pipe->unk1A -= 1024;
-
-        cos = COS_DEG(270);
-        pipe->x1 = Q_24_8_TO_INT(cos * r5) + pipe->x1;
-        pipe->x1 += r5;
-
-        sin = SIN_DEG(270);
-        pipe->y1 = Q_24_8_TO_INT(sin * r5) + pipe->y1;
-
-        value = data[++pipe->unk18].unk0;
-        if (value == (u16)-1) {
-            pipe->unk18 |= value;
-        }
-    }
-}
-
-void Handler_MusicPlant_Pipe_9(Sprite_Pipe_Horn *pipe, const Pipe_Data data[])
-{
-    s32 sin, cos;
-    s32 sin2;
-    s32 r8, r9;
-    s32 r7 = data[pipe->unk18].unk4._32;
-    u16 sinIndex = ((pipe->unk1A >> 3) + DEG_TO_SIN(315)) & ONE_CYCLE;
-
-    cos = COS_24_8(sinIndex);
-    r8 = r7 * cos;
-
-    sin = SIN_DEG(135);
-    r9 = r7 * sin;
-    pipe->x2 = Q_24_8_TO_INT(r8 - r9) + pipe->x1;
-
-    sin2 = (SIN_24_8(sinIndex) + SIN_DEG(45));
-    pipe->y2 = Q_24_8_TO_INT(sin2 * r7) + pipe->y1;
-
-    pipe->unk1A += data[pipe->unk18].unk2;
-
-    if (pipe->unk1A > ONE_CYCLE) {
-        u32 value;
-        pipe->unk1A -= 1024;
-
-        cos = SIN_DEG(90) * r7;
-        pipe->x1 = Q_24_8_TO_INT(cos - r9) + pipe->x1;
-
-        pipe->y1 = Q_24_8_TO_INT((SIN_DEG(0) + SIN_DEG(45)) * r7) + pipe->y1;
-
-        value = data[++pipe->unk18].unk0;
-        if (value == (u16)-1) {
-            pipe->unk18 |= value;
-        }
-    }
-}
-
-void Handler_MusicPlant_Pipe_10(Sprite_Pipe_Horn *pipe, const Pipe_Data data[])
-{
-    s16 sin, cos;
-    s32 r5 = data[pipe->unk18].unk4._32;
-    u16 sinIndex = ((pipe->unk1A >> 3) + 512) & ONE_CYCLE;
-
-    cos = COS_24_8(sinIndex);
-    pipe->x2 = Q_24_8_TO_INT(cos * r5) + pipe->x1 + r5;
-
-    sin = SIN_24_8(sinIndex);
-    pipe->y2 = Q_24_8_TO_INT(sin * r5) + pipe->y1;
-
-    pipe->unk1A += data[pipe->unk18].unk2;
-
-    if (pipe->unk1A > ONE_CYCLE) {
-        u32 value;
-        pipe->unk1A -= 1024;
-
-        cos = COS_24_8(640);
-        pipe->x1 = Q_24_8_TO_INT(cos * r5) + pipe->x1;
-        pipe->x1 += r5;
-
-        sin = SIN_24_8(640);
-        pipe->y1 = Q_24_8_TO_INT(sin * r5) + pipe->y1;
-
-        value = data[++pipe->unk18].unk0;
-        if (value == (u16)-1) {
-            pipe->unk18 |= value;
-        }
-    }
-}
 
 void sub_8077774(Sprite_Pipe_Horn *pipe, s32 x, s32 y)
 {
